NN-4571 refactor
