+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsmanageadjudicationsapi/services/draft/DraftAdjudicationService.kt
@@ -1,32 +1,18 @@
-package uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services
+package uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.draft
 
 import org.springframework.stereotype.Service
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.DamageRequestItem
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.EvidenceRequestItem
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.IncidentRoleAssociatedPrisonerRequest
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.IncidentRoleRequest
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.OffenceDetailsRequestItem
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.WitnessRequestItem
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.DamageDto
+import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.draft.IncidentRoleAssociatedPrisonerRequest
+import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.draft.IncidentRoleRequest
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.DraftAdjudicationDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.EvidenceDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.IncidentDetailsDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.IncidentRoleDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.IncidentStatementDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.OffenceDetailsDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.OffenceRuleDetailsDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.WitnessDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.Damage
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.DraftAdjudication
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.Evidence
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.IncidentDetails
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.IncidentRole
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.IncidentStatement
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.Offence
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.Witness
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.repositories.DraftAdjudicationRepository
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.security.AuthenticationFacade
+import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.IncidentRoleRuleLookup
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.IncidentRoleRuleLookup.Companion.associatedPrisonerInformationRequired
+import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.OffenceCodeLookupService
 import java.time.LocalDateTime
 import javax.persistence.EntityNotFoundException
 import javax.transaction.Transactional
@@ -126,28 +112,6 @@ class DraftAdjudicationService(
   }
 
   fun getDraftAdjudicationDetails(id: Long): DraftAdjudicationDto = findToDto(id)
-
-  fun setOffenceDetails(id: Long, offenceDetails: List<OffenceDetailsRequestItem>): DraftAdjudicationDto {
-    throwIfEmpty(offenceDetails)
-
-    val draftAdjudication = draftAdjudicationRepository.findById(id).orElseThrow { throwEntityNotFoundException(id) }
-    // NOTE: new flow sets isYouthOffender first, therefore if we do not have this set we must throw as .Dto requires it
-    ValidationChecks.APPLICABLE_RULES.validate(draftAdjudication)
-
-    val newValuesToStore = offenceDetails.map {
-      Offence(
-        offenceCode = it.offenceCode,
-        victimPrisonersNumber = it.victimPrisonersNumber?.ifBlank { null },
-        victimStaffUsername = it.victimStaffUsername?.ifBlank { null },
-        victimOtherPersonsName = it.victimOtherPersonsName?.ifBlank { null },
-      )
-    }.toMutableList()
-
-    draftAdjudication.offenceDetails.clear()
-    draftAdjudication.offenceDetails.addAll(newValuesToStore)
-
-    return saveToDto(draftAdjudication)
-  }
 
   fun addIncidentStatement(id: Long, statement: String?, completed: Boolean?): DraftAdjudicationDto {
     throwIfStatementAndCompletedIsNull(statement, completed)
@@ -264,65 +228,6 @@ class DraftAdjudicationService(
     return saveToDto(draftAdjudication)
   }
 
-  fun setDamages(id: Long, damages: List<DamageRequestItem>): DraftAdjudicationDto {
-    val draftAdjudication = draftAdjudicationRepository.findById(id).orElseThrow { throwEntityNotFoundException(id) }
-    val reporter = authenticationFacade.currentUsername!!
-
-    draftAdjudication.damages.clear()
-    draftAdjudication.damages.addAll(
-      damages.map {
-        Damage(
-          code = it.code,
-          details = it.details,
-          reporter = reporter
-        )
-      }
-    )
-    draftAdjudication.damagesSaved = true
-
-    return saveToDto(draftAdjudication)
-  }
-
-  fun setEvidence(id: Long, evidence: List<EvidenceRequestItem>): DraftAdjudicationDto {
-    val draftAdjudication = draftAdjudicationRepository.findById(id).orElseThrow { throwEntityNotFoundException(id) }
-    val reporter = authenticationFacade.currentUsername!!
-
-    draftAdjudication.evidence.clear()
-    draftAdjudication.evidence.addAll(
-      evidence.map {
-        Evidence(
-          code = it.code,
-          identifier = it.identifier,
-          details = it.details,
-          reporter = reporter
-        )
-      }
-    )
-    draftAdjudication.evidenceSaved = true
-
-    return saveToDto(draftAdjudication)
-  }
-
-  fun setWitnesses(id: Long, witnesses: List<WitnessRequestItem>): DraftAdjudicationDto {
-    val draftAdjudication = draftAdjudicationRepository.findById(id).orElseThrow { throwEntityNotFoundException(id) }
-    val reporter = authenticationFacade.currentUsername!!
-
-    draftAdjudication.witnesses.clear()
-    draftAdjudication.witnesses.addAll(
-      witnesses.map {
-        Witness(
-          code = it.code,
-          firstName = it.firstName,
-          lastName = it.lastName,
-          reporter = reporter
-        )
-      }
-    )
-    draftAdjudication.witnessesSaved = true
-
-    return saveToDto(draftAdjudication)
-  }
-
   fun getCurrentUsersInProgressDraftAdjudications(agencyId: String): List<DraftAdjudicationDto> {
     val username = authenticationFacade.currentUsername ?: return emptyList()
 
@@ -334,70 +239,8 @@ class DraftAdjudicationService(
       .map { it.toDto() }
   }
 
-  fun lookupRuleDetails(offenceCode: Int, isYouthOffender: Boolean): OffenceRuleDetailsDto {
-    return OffenceRuleDetailsDto(
-      paragraphNumber = offenceCodeLookupService.getParagraphNumber(offenceCode, isYouthOffender),
-      paragraphDescription = offenceCodeLookupService.getParagraphDescription(offenceCode, isYouthOffender),
-    )
-  }
-
   private fun throwIfStatementAndCompletedIsNull(statement: String?, completed: Boolean?) {
     if (statement == null && completed == null)
       throw IllegalArgumentException("Please supply either a statement or the completed value")
   }
-
-  private fun throwIfEmpty(toTest: List<Any>) {
-    if (toTest.isEmpty())
-      throw IllegalArgumentException("Please supply at least one set of items")
-  }
 }
-
-fun IncidentDetails.toDto(): IncidentDetailsDto = IncidentDetailsDto(
-  locationId = this.locationId,
-  dateTimeOfIncident = this.dateTimeOfIncident,
-  dateTimeOfDiscovery = this.dateTimeOfDiscovery,
-  handoverDeadline = this.handoverDeadline,
-)
-
-fun IncidentRole.toDto(isYouthOffender: Boolean): IncidentRoleDto = IncidentRoleDto(
-  roleCode = this.roleCode,
-  offenceRule = IncidentRoleRuleLookup.getOffenceRuleDetails(this.roleCode, isYouthOffender),
-  associatedPrisonersNumber = this.associatedPrisonersNumber,
-  associatedPrisonersName = this.associatedPrisonersName,
-)
-
-fun Offence.toDto(offenceCodeLookupService: OffenceCodeLookupService, isYouthOffender: Boolean): OffenceDetailsDto = OffenceDetailsDto(
-  offenceCode = this.offenceCode,
-  offenceRule = OffenceRuleDetailsDto(
-    paragraphNumber = offenceCodeLookupService.getParagraphNumber(offenceCode, isYouthOffender),
-    paragraphDescription = offenceCodeLookupService.getParagraphDescription(offenceCode, isYouthOffender),
-  ),
-  victimPrisonersNumber = this.victimPrisonersNumber,
-  victimStaffUsername = this.victimStaffUsername,
-  victimOtherPersonsName = this.victimOtherPersonsName,
-)
-
-fun IncidentStatement.toDto(): IncidentStatementDto = IncidentStatementDto(
-  statement = this.statement!!,
-  completed = this.completed,
-)
-
-fun Damage.toDto(): DamageDto = DamageDto(
-  code = this.code,
-  details = this.details,
-  reporter = this.reporter
-)
-
-fun Evidence.toDto(): EvidenceDto = EvidenceDto(
-  code = this.code,
-  identifier = this.identifier,
-  details = this.details,
-  reporter = this.reporter
-)
-
-fun Witness.toDto(): WitnessDto = WitnessDto(
-  code = this.code,
-  firstName = this.firstName,
-  lastName = this.lastName,
-  reporter = this.reporter
-)
