+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsmanageadjudicationsapi/services/draft/DraftAdjudicationServiceTest.kt
@@ -1,6 +1,5 @@
-package uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services
+package uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.draft
 
-import com.microsoft.applicationinsights.TelemetryClient
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.assertj.core.api.Java6Assertions.assertThat
 import org.assertj.core.groups.Tuple
@@ -11,102 +10,35 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 import org.mockito.ArgumentCaptor
 import org.mockito.kotlin.any
-import org.mockito.kotlin.mock
-import org.mockito.kotlin.times
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.DamageRequestItem
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.EvidenceRequestItem
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.IncidentRoleAssociatedPrisonerRequest
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.IncidentRoleRequest
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.OffenceDetailsRequestItem
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.WitnessRequestItem
+import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.draft.IncidentRoleAssociatedPrisonerRequest
+import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.draft.IncidentRoleRequest
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.DraftAdjudicationDto
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.IncidentDetailsDto
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.IncidentRoleDto
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.IncidentStatementDto
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.OffenceDetailsDto
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.OffenceRuleDetailsDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.Damage
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.DamageCode
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.DraftAdjudication
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.Evidence
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.EvidenceCode
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.IncidentDetails
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.IncidentRole
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.IncidentStatement
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.Offence
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.Witness
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.WitnessCode
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.gateways.PrisonApiGateway
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.repositories.DraftAdjudicationRepository
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.repositories.ReportedAdjudicationRepository
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.security.AuthenticationFacade
+import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.IncidentRoleRuleLookup
 import java.time.Clock
-import java.time.Instant.ofEpochMilli
 import java.time.LocalDateTime
-import java.time.ZoneId
 import java.util.Optional
 import java.util.function.Supplier
 import javax.persistence.EntityNotFoundException
 
-class DraftAdjudicationServiceTest {
-  private val draftAdjudicationRepository: DraftAdjudicationRepository = mock()
-  private val reportedAdjudicationRepository: ReportedAdjudicationRepository = mock()
-  private val prisonApiGateway: PrisonApiGateway = mock()
-  private val offenceCodeLookupService: OffenceCodeLookupService = mock()
-  private val authenticationFacade: AuthenticationFacade = mock()
-  private val telemetryClient: TelemetryClient = mock()
-  private val clock: Clock = Clock.fixed(ofEpochMilli(0), ZoneId.systemDefault())
+class DraftAdjudicationServiceTest : DraftAdjudicationTestBase() {
 
-  private lateinit var draftAdjudicationService: DraftAdjudicationService
-
-  @BeforeEach
-  fun beforeEach() {
-    draftAdjudicationService =
-      DraftAdjudicationService(
-        draftAdjudicationRepository,
-        offenceCodeLookupService,
-        authenticationFacade,
-      )
-
-    // Set up offence code mocks
-    whenever(offenceCodeLookupService.getCommittedOnOwnNomisOffenceCodes(2, false)).thenReturn(
-      OFFENCE_CODE_2_NOMIS_CODE_ON_OWN
+  private val draftAdjudicationService =
+    DraftAdjudicationService(
+      draftAdjudicationRepository,
+      offenceCodeLookupService,
+      authenticationFacade,
     )
-    whenever(offenceCodeLookupService.getNotCommittedOnOwnNomisOffenceCode(2, false)).thenReturn(
-      OFFENCE_CODE_2_NOMIS_CODE_ASSISTED
-    )
-    whenever(offenceCodeLookupService.getParagraphNumber(2, false)).thenReturn(OFFENCE_CODE_2_PARAGRAPH_NUMBER)
-    whenever(offenceCodeLookupService.getParagraphDescription(2, false)).thenReturn(OFFENCE_CODE_2_PARAGRAPH_DESCRIPTION)
-
-    whenever(offenceCodeLookupService.getCommittedOnOwnNomisOffenceCodes(3, false)).thenReturn(
-      OFFENCE_CODE_3_NOMIS_CODE_ON_OWN
-    )
-    whenever(offenceCodeLookupService.getNotCommittedOnOwnNomisOffenceCode(3, false)).thenReturn(
-      OFFENCE_CODE_3_NOMIS_CODE_ASSISTED
-    )
-    whenever(offenceCodeLookupService.getParagraphNumber(3, false)).thenReturn(OFFENCE_CODE_3_PARAGRAPH_NUMBER)
-    whenever(offenceCodeLookupService.getParagraphDescription(3, false)).thenReturn(OFFENCE_CODE_3_PARAGRAPH_DESCRIPTION)
-
-    whenever(offenceCodeLookupService.getCommittedOnOwnNomisOffenceCodes(2, true)).thenReturn(
-      YOUTH_OFFENCE_CODE_2_NOMIS_CODE_ON_OWN
-    )
-    whenever(offenceCodeLookupService.getNotCommittedOnOwnNomisOffenceCode(2, true)).thenReturn(
-      YOUTH_OFFENCE_CODE_2_NOMIS_CODE_ASSISTED
-    )
-    whenever(offenceCodeLookupService.getParagraphNumber(2, true)).thenReturn(YOUTH_OFFENCE_CODE_2_PARAGRAPH_NUMBER)
-    whenever(offenceCodeLookupService.getParagraphDescription(2, true)).thenReturn(YOUTH_OFFENCE_CODE_2_PARAGRAPH_DESCRIPTION)
-
-    whenever(offenceCodeLookupService.getCommittedOnOwnNomisOffenceCodes(3, true)).thenReturn(
-      YOUTH_OFFENCE_CODE_3_NOMIS_CODE_ON_OWN
-    )
-    whenever(offenceCodeLookupService.getNotCommittedOnOwnNomisOffenceCode(3, true)).thenReturn(
-      YOUTH_OFFENCE_CODE_3_NOMIS_CODE_ASSISTED
-    )
-    whenever(offenceCodeLookupService.getParagraphNumber(3, true)).thenReturn(YOUTH_OFFENCE_CODE_3_PARAGRAPH_NUMBER)
-    whenever(offenceCodeLookupService.getParagraphDescription(3, true)).thenReturn(YOUTH_OFFENCE_CODE_3_PARAGRAPH_DESCRIPTION)
-  }
 
   @Nested
   inner class StartDraftAdjudications {
@@ -147,15 +79,6 @@ class DraftAdjudicationServiceTest {
       val argumentCaptor = ArgumentCaptor.forClass(DraftAdjudication::class.java)
 
       verify(draftAdjudicationRepository).save(argumentCaptor.capture())
-      verify(telemetryClient).trackEvent(
-        DraftAdjudicationService.TELEMETRY_EVENT,
-        mapOf(
-          "adjudicationNumber" to draftAdjudication.id.toString(),
-          "agencyId" to "MDI",
-          "reportNumber" to "null"
-        ),
-        null
-      )
 
       assertThat(draftAdjudication)
         .extracting("id", "prisonerNumber")
@@ -172,15 +95,6 @@ class DraftAdjudicationServiceTest {
 
   @Nested
   inner class DraftAdjudicationDetails {
-    @Test
-    fun `throws an entity not found if the draft adjudication for the supplied id does not exists`() {
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.empty())
-
-      assertThatThrownBy {
-        draftAdjudicationService.getDraftAdjudicationDetails(1)
-      }.isInstanceOf(EntityNotFoundException::class.java)
-        .hasMessageContaining("DraftAdjudication not found for 1")
-    }
 
     @Test
     fun `returns the draft adjudication`() {
@@ -330,16 +244,6 @@ class DraftAdjudicationServiceTest {
       )
 
     @Test
-    fun `throws an entity not found if the draft adjudication for the supplied id does not exists`() {
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.empty())
-
-      assertThatThrownBy {
-        draftAdjudicationService.editIncidentRole(1, IncidentRoleRequest("1"), false)
-      }.isInstanceOf(EntityNotFoundException::class.java)
-        .hasMessageContaining("DraftAdjudication not found for 1")
-    }
-
-    @Test
     fun `throws state exception if isYouthOffender is not set`() {
       whenever(draftAdjudicationRepository.findById(any())).thenReturn(
         Optional.of(
@@ -462,19 +366,6 @@ class DraftAdjudicationServiceTest {
       )
 
     @Test
-    fun `throws an entity not found if the draft adjudication for the supplied id does not exists`() {
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.empty())
-
-      assertThatThrownBy {
-        draftAdjudicationService.setIncidentRoleAssociatedPrisoner(
-          1,
-          IncidentRoleAssociatedPrisonerRequest("A1234AA", "A name")
-        )
-      }.isInstanceOf(EntityNotFoundException::class.java)
-        .hasMessageContaining("DraftAdjudication not found for 1")
-    }
-
-    @Test
     fun `throws state exception if incident role is not set`() {
       whenever(draftAdjudicationRepository.findById(any())).thenReturn(
         Optional.of(
@@ -533,199 +424,7 @@ class DraftAdjudicationServiceTest {
   }
 
   @Nested
-  inner class SetOffenceDetails {
-    private val draftAdjudicationEntity = DraftAdjudication(
-      id = 1,
-      prisonerNumber = "A12345",
-      agencyId = "MDI",
-      incidentDetails = incidentDetails(2L, clock),
-      incidentRole = incidentRoleWithNoValuesSet(),
-    )
-
-    @Test
-    fun `throws an entity not found if the draft adjudication for the supplied id does not exists`() {
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.empty())
-
-      assertThatThrownBy {
-        draftAdjudicationService.setOffenceDetails(1, listOf(BASIC_OFFENCE_DETAILS_REQUEST))
-      }.isInstanceOf(EntityNotFoundException::class.java)
-        .hasMessageContaining("DraftAdjudication not found for 1")
-    }
-
-    @Test
-    fun `throws state exception if isYouthOffender is not set`() {
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.of(draftAdjudicationEntity))
-
-      assertThatThrownBy {
-        draftAdjudicationService.setOffenceDetails(1, listOf(BASIC_OFFENCE_DETAILS_REQUEST))
-      }.isInstanceOf(IllegalStateException::class.java)
-        .hasMessageContaining(ValidationChecks.APPLICABLE_RULES.errorMessage)
-    }
-
-    @ParameterizedTest
-    @CsvSource(
-      "true",
-      "false",
-    )
-    fun `adds the offence details to a draft adjudication`(
-      isYouthOffender: Boolean,
-    ) {
-      var offenceDetailsToAdd = listOf(BASIC_OFFENCE_DETAILS_REQUEST, FULL_OFFENCE_DETAILS_REQUEST)
-      var offenceDetailsToSave = mutableListOf(BASIC_OFFENCE_DETAILS_DB_ENTITY, FULL_OFFENCE_DETAILS_DB_ENTITY)
-      var expectedOffenceDetailsResponse = listOf(BASIC_OFFENCE_DETAILS_RESPONSE_DTO, FULL_OFFENCE_DETAILS_RESPONSE_DTO)
-      if (isYouthOffender) {
-        offenceDetailsToAdd = listOf(YOUTH_OFFENCE_DETAILS_REQUEST)
-        offenceDetailsToSave = mutableListOf(YOUTH_OFFENCE_DETAILS_DB_ENTITY)
-        expectedOffenceDetailsResponse = listOf(YOUTH_OFFENCE_DETAILS_RESPONSE_DTO)
-      }
-
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(
-        Optional.of(
-          draftAdjudicationEntity.also {
-            it.isYouthOffender = isYouthOffender
-          }
-        )
-      )
-
-      whenever(draftAdjudicationRepository.save(any())).thenReturn(
-        draftAdjudicationEntity.copy(
-          offenceDetails = offenceDetailsToSave
-        )
-      )
-
-      val draftAdjudication = draftAdjudicationService.setOffenceDetails(1, offenceDetailsToAdd)
-
-      assertThat(draftAdjudication)
-        .extracting("id", "prisonerNumber")
-        .contains(1L, "A12345")
-
-      assertThat(draftAdjudication.offenceDetails).isEqualTo(expectedOffenceDetailsResponse)
-
-      val argumentCaptor = ArgumentCaptor.forClass(DraftAdjudication::class.java)
-      verify(draftAdjudicationRepository).save(argumentCaptor.capture())
-
-      assertThat(argumentCaptor.value.offenceDetails).isEqualTo(offenceDetailsToSave)
-    }
-
-    @Test
-    fun `edits the offence details of an existing draft adjudication`() {
-      val existingOffenceDetails = mutableListOf(Offence(offenceCode = 1))
-      val offenceDetailsToUse = listOf(BASIC_OFFENCE_DETAILS_REQUEST, FULL_OFFENCE_DETAILS_REQUEST)
-      val offenceDetailsToSave = mutableListOf(BASIC_OFFENCE_DETAILS_DB_ENTITY, FULL_OFFENCE_DETAILS_DB_ENTITY)
-      val expectedOffenceDetailsResponse = listOf(BASIC_OFFENCE_DETAILS_RESPONSE_DTO, FULL_OFFENCE_DETAILS_RESPONSE_DTO)
-      val existingDraftAdjudicationEntity = DraftAdjudication(
-        id = 1,
-        prisonerNumber = "A12345",
-        agencyId = "MDI",
-        incidentDetails = IncidentDetails(
-          locationId = 1,
-          dateTimeOfIncident = LocalDateTime.now(clock),
-          dateTimeOfDiscovery = LocalDateTime.now(clock).plusDays(1),
-          handoverDeadline = DATE_TIME_DRAFT_ADJUDICATION_HANDOVER_DEADLINE
-        ),
-        incidentRole = incidentRoleWithNoValuesSet(),
-        offenceDetails = existingOffenceDetails,
-        isYouthOffender = false
-      )
-
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.of(existingDraftAdjudicationEntity))
-
-      whenever(draftAdjudicationRepository.save(any())).thenReturn(
-        existingDraftAdjudicationEntity.copy(
-          offenceDetails = offenceDetailsToSave
-        )
-      )
-
-      val draftAdjudication = draftAdjudicationService.setOffenceDetails(1, offenceDetailsToUse)
-
-      assertThat(draftAdjudication)
-        .extracting("id", "prisonerNumber")
-        .contains(1L, "A12345")
-
-      assertThat(draftAdjudication.offenceDetails).isEqualTo(expectedOffenceDetailsResponse)
-
-      val argumentCaptor = ArgumentCaptor.forClass(DraftAdjudication::class.java)
-      verify(draftAdjudicationRepository).save(argumentCaptor.capture())
-
-      assertThat(argumentCaptor.value.offenceDetails).isEqualTo(offenceDetailsToSave)
-    }
-
-    @Test
-    fun `treats empty strings as null values`() {
-      val offenceDetailsToAdd = listOf(
-        OffenceDetailsRequestItem(
-          offenceCode = 2,
-          victimPrisonersNumber = "",
-          victimStaffUsername = "",
-          victimOtherPersonsName = "",
-        )
-      )
-      val offenceDetailsToSave = mutableListOf(
-        Offence(
-          offenceCode = 2,
-        )
-      )
-      val expectedOffenceDetailsResponse = listOf(
-        OffenceDetailsDto(
-          offenceCode = 2,
-          offenceRule = OffenceRuleDetailsDto(
-            paragraphNumber = OFFENCE_CODE_2_PARAGRAPH_NUMBER,
-            paragraphDescription = OFFENCE_CODE_2_PARAGRAPH_DESCRIPTION,
-          ),
-        )
-      )
-
-      val draftAdjudicationEntity = DraftAdjudication(
-        id = 1,
-        prisonerNumber = "A12345",
-        agencyId = "MDI",
-        incidentDetails = incidentDetails(2L, clock),
-        incidentRole = incidentRoleWithNoValuesSet(),
-        isYouthOffender = false
-      )
-
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.of(draftAdjudicationEntity))
-
-      whenever(draftAdjudicationRepository.save(any())).thenReturn(
-        draftAdjudicationEntity.copy(
-          offenceDetails = offenceDetailsToSave
-        )
-      )
-
-      val draftAdjudication = draftAdjudicationService.setOffenceDetails(1, offenceDetailsToAdd)
-
-      assertThat(draftAdjudication)
-        .extracting("id", "prisonerNumber")
-        .contains(1L, "A12345")
-
-      assertThat(draftAdjudication.offenceDetails).isEqualTo(expectedOffenceDetailsResponse)
-
-      val argumentCaptor = ArgumentCaptor.forClass(DraftAdjudication::class.java)
-      verify(draftAdjudicationRepository).save(argumentCaptor.capture())
-
-      assertThat(argumentCaptor.value.offenceDetails).isEqualTo(offenceDetailsToSave)
-    }
-
-    @Test
-    fun `throws an IllegalArgumentException when no offence details are provided`() {
-      assertThatThrownBy {
-        draftAdjudicationService.setOffenceDetails(1, listOf())
-      }.isInstanceOf(IllegalArgumentException::class.java)
-        .hasMessageContaining("Please supply at least one set of items")
-    }
-  }
-
-  @Nested
   inner class AddIncidentStatement {
-    @Test
-    fun `throws an entity not found if the draft adjudication for the supplied id does not exists`() {
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.empty())
-
-      assertThatThrownBy {
-        draftAdjudicationService.addIncidentStatement(1, "test", false)
-      }.isInstanceOf(EntityNotFoundException::class.java)
-        .hasMessageContaining("DraftAdjudication not found for 1")
-    }
 
     @Test
     fun `adds an incident statement to a draft adjudication`() {
@@ -794,21 +493,6 @@ class DraftAdjudicationServiceTest {
 
   @Nested
   inner class EditIncidentDetails {
-    @Test
-    fun `throws an entity not found if the draft adjudication for the supplied id does not exists`() {
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.empty())
-
-      assertThatThrownBy {
-        draftAdjudicationService.editIncidentDetails(
-          1,
-          2,
-          DATE_TIME_OF_INCIDENT,
-          DATE_TIME_OF_INCIDENT.plusDays(1),
-        )
-      }.isInstanceOf(EntityNotFoundException::class.java)
-        .hasMessageContaining("DraftAdjudication not found for 1")
-    }
-
     @Test
     fun `throws exception if date of discovery before incident date`() {
       assertThatThrownBy {
@@ -890,16 +574,6 @@ class DraftAdjudicationServiceTest {
 
   @Nested
   inner class EditIncidentStatement {
-    @Test
-    fun `throws an entity not found if the draft adjudication for the supplied id does not exists`() {
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.empty())
-
-      assertThatThrownBy {
-        draftAdjudicationService.editIncidentStatement(1, "new statement", false)
-      }.isInstanceOf(EntityNotFoundException::class.java)
-        .hasMessageContaining("DraftAdjudication not found for 1")
-    }
-
     @Test
     fun `throws an entity not found if the incident statement does not exist`() {
       whenever(draftAdjudicationRepository.findById(any())).thenReturn(
@@ -1082,16 +756,6 @@ class DraftAdjudicationServiceTest {
         isYouthOffender = true
       )
 
-    @Test
-    fun `throws an entity not found if the draft adjudication for the supplied id does not exists`() {
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.empty())
-
-      assertThatThrownBy {
-        draftAdjudicationService.setIncidentApplicableRule(1, false, true)
-      }.isInstanceOf(EntityNotFoundException::class.java)
-        .hasMessageContaining("DraftAdjudication not found for 1")
-    }
-
     @ParameterizedTest
     @CsvSource("true", "false")
     fun `saves incident applicable rule`(deleteOffences: Boolean) {
@@ -1114,226 +778,7 @@ class DraftAdjudicationServiceTest {
     }
   }
 
-  @Nested
-  inner class AddDamages {
-    private val draftAdjudication =
-      DraftAdjudication(
-        id = 1,
-        prisonerNumber = "A12345",
-        agencyId = "MDI",
-        incidentDetails = incidentDetails(2L, now),
-        incidentRole = incidentRoleWithAllValuesSet(),
-        incidentStatement = IncidentStatement(
-          statement = "Example statement",
-          completed = false
-        ),
-        offenceDetails = mutableListOf(BASIC_OFFENCE_DETAILS_DB_ENTITY, FULL_OFFENCE_DETAILS_DB_ENTITY),
-        isYouthOffender = true,
-        damages = mutableListOf(
-          Damage(code = DamageCode.CLEANING, details = "details", reporter = "Fred")
-        )
-      )
-
-    @BeforeEach
-    fun init() {
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.of(draftAdjudication))
-      whenever(draftAdjudicationRepository.save(any())).thenReturn(draftAdjudication)
-      whenever(authenticationFacade.currentUsername).thenReturn("Fred")
-    }
-
-    @Test
-    fun `add damages to adjudication`() {
-      val response = draftAdjudicationService.setDamages(
-        1,
-        listOf(
-          DamageRequestItem(DamageCode.CLEANING, "details")
-        )
-      )
-
-      val argumentCaptor = ArgumentCaptor.forClass(DraftAdjudication::class.java)
-      verify(draftAdjudicationRepository).save(argumentCaptor.capture())
-
-      assertThat(argumentCaptor.value.damages.size).isEqualTo(1)
-      assertThat(argumentCaptor.value.damages.first().code).isEqualTo(DamageCode.CLEANING)
-      assertThat(argumentCaptor.value.damages.first().details).isEqualTo("details")
-      assertThat(argumentCaptor.value.damages.first().reporter).isEqualTo("Fred")
-      assertThat(argumentCaptor.value.damagesSaved).isEqualTo(true)
-
-      assertThat(response).isNotNull
-    }
-
-    @Test
-    fun `throws an entity not found if the draft adjudication for the supplied id does not exists`() {
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.empty())
-
-      assertThatThrownBy {
-        draftAdjudicationService.setDamages(1, listOf(DamageRequestItem(DamageCode.CLEANING, "details")))
-      }.isInstanceOf(EntityNotFoundException::class.java)
-        .hasMessageContaining("DraftAdjudication not found for 1")
-    }
-  }
-
-  @Nested
-  inner class AddEvidence {
-    private val draftAdjudication =
-      DraftAdjudication(
-        id = 1,
-        prisonerNumber = "A12345",
-        agencyId = "MDI",
-        incidentDetails = incidentDetails(2L, now),
-        incidentRole = incidentRoleWithAllValuesSet(),
-        incidentStatement = IncidentStatement(
-          statement = "Example statement",
-          completed = false
-        ),
-        offenceDetails = mutableListOf(BASIC_OFFENCE_DETAILS_DB_ENTITY, FULL_OFFENCE_DETAILS_DB_ENTITY),
-        isYouthOffender = true,
-        evidence = mutableListOf(
-          Evidence(code = EvidenceCode.PHOTO, details = "details", reporter = "Fred")
-        )
-      )
-
-    @BeforeEach
-    fun init() {
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.of(draftAdjudication))
-      whenever(draftAdjudicationRepository.save(any())).thenReturn(draftAdjudication)
-      whenever(authenticationFacade.currentUsername).thenReturn("Fred")
-    }
-
-    @Test
-    fun `add evidence to adjudication`() {
-      val response = draftAdjudicationService.setEvidence(
-        1,
-        listOf(
-          EvidenceRequestItem(code = EvidenceCode.PHOTO, details = "details")
-        )
-      )
-
-      val argumentCaptor = ArgumentCaptor.forClass(DraftAdjudication::class.java)
-      verify(draftAdjudicationRepository).save(argumentCaptor.capture())
-
-      assertThat(argumentCaptor.value.evidence.size).isEqualTo(1)
-      assertThat(argumentCaptor.value.evidence.first().code).isEqualTo(EvidenceCode.PHOTO)
-      assertThat(argumentCaptor.value.evidence.first().details).isEqualTo("details")
-      assertThat(argumentCaptor.value.evidence.first().reporter).isEqualTo("Fred")
-      assertThat(argumentCaptor.value.evidenceSaved).isEqualTo(true)
-
-      assertThat(response).isNotNull
-    }
-
-    @Test
-    fun `throws an entity not found if the draft adjudication for the supplied id does not exists`() {
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.empty())
-
-      assertThatThrownBy {
-        draftAdjudicationService.setEvidence(1, listOf(EvidenceRequestItem(code = EvidenceCode.PHOTO, details = "details")))
-      }.isInstanceOf(EntityNotFoundException::class.java)
-        .hasMessageContaining("DraftAdjudication not found for 1")
-    }
-  }
-
-  @Nested
-  inner class AddWitnesses {
-    private val draftAdjudication =
-      DraftAdjudication(
-        id = 1,
-        prisonerNumber = "A12345",
-        agencyId = "MDI",
-        incidentDetails = incidentDetails(2L, now),
-        incidentRole = incidentRoleWithAllValuesSet(),
-        incidentStatement = IncidentStatement(
-          statement = "Example statement",
-          completed = false
-        ),
-        offenceDetails = mutableListOf(BASIC_OFFENCE_DETAILS_DB_ENTITY, FULL_OFFENCE_DETAILS_DB_ENTITY),
-        isYouthOffender = true,
-        witnesses = mutableListOf(
-          Witness(code = WitnessCode.OFFICER, firstName = "prison", lastName = "officer", reporter = "Fred")
-        )
-      )
-
-    @BeforeEach
-    fun init() {
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.of(draftAdjudication))
-      whenever(draftAdjudicationRepository.save(any())).thenReturn(draftAdjudication)
-      whenever(authenticationFacade.currentUsername).thenReturn("Fred")
-    }
-
-    @Test
-    fun `add witnesses to adjudication`() {
-      val response = draftAdjudicationService.setWitnesses(
-        1,
-        listOf(
-          WitnessRequestItem(code = WitnessCode.OFFICER, firstName = "prison", lastName = "officer")
-        )
-      )
-
-      val argumentCaptor = ArgumentCaptor.forClass(DraftAdjudication::class.java)
-      verify(draftAdjudicationRepository).save(argumentCaptor.capture())
-
-      assertThat(argumentCaptor.value.witnesses.size).isEqualTo(1)
-      assertThat(argumentCaptor.value.witnesses.first().code).isEqualTo(WitnessCode.OFFICER)
-      assertThat(argumentCaptor.value.witnesses.first().firstName).isEqualTo("prison")
-      assertThat(argumentCaptor.value.witnesses.first().lastName).isEqualTo("officer")
-      assertThat(argumentCaptor.value.witnesses.first().reporter).isEqualTo("Fred")
-      assertThat(argumentCaptor.value.witnessesSaved).isEqualTo(true)
-
-      assertThat(response).isNotNull
-    }
-
-    @Test
-    fun `throws an entity not found if the draft adjudication for the supplied id does not exists`() {
-      whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.empty())
-
-      assertThatThrownBy {
-        draftAdjudicationService.setWitnesses(1, listOf(WitnessRequestItem(code = WitnessCode.OFFICER, firstName = "prison", lastName = "officer")))
-      }.isInstanceOf(EntityNotFoundException::class.java)
-        .hasMessageContaining("DraftAdjudication not found for 1")
-    }
-  }
-
   companion object {
-    val now = LocalDateTime.now()
-
-    private val DATE_TIME_OF_INCIDENT = LocalDateTime.of(2010, 10, 12, 10, 0)
-    private val DATE_TIME_DRAFT_ADJUDICATION_HANDOVER_DEADLINE = LocalDateTime.of(2010, 10, 14, 10, 0)
-    private val REPORTED_DATE_TIME = DATE_TIME_OF_INCIDENT.plusDays(1)
-
-    private val INCIDENT_ROLE_CODE = "25a"
-    private val INCIDENT_ROLE_PARAGRAPH_NUMBER = "25(a)"
-    private val INCIDENT_ROLE_PARAGRAPH_DESCRIPTION = "Attempts to commit any of the foregoing offences:"
-    private val INCIDENT_ROLE_ASSOCIATED_PRISONERS_NUMBER = "B23456"
-    private val INCIDENT_ROLE_ASSOCIATED_PRISONERS_NAME = "Associated Prisoner"
-
-    private const val OFFENCE_CODE_2_NOMIS_CODE_ON_OWN = "5b"
-    private const val OFFENCE_CODE_2_NOMIS_CODE_ASSISTED = "25z"
-    private const val OFFENCE_CODE_2_PARAGRAPH_NUMBER = "5(b)"
-    private const val OFFENCE_CODE_2_PARAGRAPH_DESCRIPTION = "A paragraph description"
-    private const val OFFENCE_CODE_3_NOMIS_CODE_ON_OWN = "5f"
-    private const val OFFENCE_CODE_3_NOMIS_CODE_ASSISTED = "25f"
-    private const val OFFENCE_CODE_3_PARAGRAPH_NUMBER = "6(a)"
-    private const val OFFENCE_CODE_3_PARAGRAPH_DESCRIPTION = "Another paragraph description"
-
-    private const val YOUTH_OFFENCE_CODE_2_PARAGRAPH_NUMBER = "7(b)"
-    private const val YOUTH_OFFENCE_CODE_2_PARAGRAPH_DESCRIPTION = "A youth paragraph description"
-    private const val YOUTH_OFFENCE_CODE_2_NOMIS_CODE_ON_OWN = "7b"
-    private const val YOUTH_OFFENCE_CODE_2_NOMIS_CODE_ASSISTED = "29z"
-    private const val YOUTH_OFFENCE_CODE_3_PARAGRAPH_NUMBER = "17(b)"
-    private const val YOUTH_OFFENCE_CODE_3_PARAGRAPH_DESCRIPTION = "Another youth paragraph description"
-    private const val YOUTH_OFFENCE_CODE_3_NOMIS_CODE_ON_OWN = "17b"
-    private const val YOUTH_OFFENCE_CODE_3_NOMIS_CODE_ASSISTED = "29f"
-
-    private val BASIC_OFFENCE_DETAILS_REQUEST = OffenceDetailsRequestItem(offenceCode = 2)
-    private val BASIC_OFFENCE_DETAILS_RESPONSE_DTO = OffenceDetailsDto(
-      offenceCode = BASIC_OFFENCE_DETAILS_REQUEST.offenceCode,
-      offenceRule = OffenceRuleDetailsDto(
-        paragraphNumber = OFFENCE_CODE_2_PARAGRAPH_NUMBER,
-        paragraphDescription = OFFENCE_CODE_2_PARAGRAPH_DESCRIPTION,
-      )
-    )
-    private val BASIC_OFFENCE_DETAILS_DB_ENTITY = Offence(
-      offenceCode = BASIC_OFFENCE_DETAILS_RESPONSE_DTO.offenceCode,
-    )
 
     fun incidentDetails(locationId: Long, clock: Clock) = IncidentDetails(
       locationId = locationId,
@@ -1347,41 +792,6 @@ class DraftAdjudicationServiceTest {
       dateTimeOfIncident = now,
       dateTimeOfDiscovery = now.plusDays(1),
       handoverDeadline = DATE_TIME_DRAFT_ADJUDICATION_HANDOVER_DEADLINE
-    )
-
-    private val FULL_OFFENCE_DETAILS_REQUEST = OffenceDetailsRequestItem(
-      offenceCode = 3,
-      victimPrisonersNumber = "A1234AA",
-      victimStaffUsername = "ABC12D",
-      victimOtherPersonsName = "A name",
-    )
-    private val FULL_OFFENCE_DETAILS_RESPONSE_DTO = OffenceDetailsDto(
-      offenceCode = FULL_OFFENCE_DETAILS_REQUEST.offenceCode,
-      offenceRule = OffenceRuleDetailsDto(
-        paragraphNumber = OFFENCE_CODE_3_PARAGRAPH_NUMBER,
-        paragraphDescription = OFFENCE_CODE_3_PARAGRAPH_DESCRIPTION,
-      ),
-      victimPrisonersNumber = FULL_OFFENCE_DETAILS_REQUEST.victimPrisonersNumber,
-      victimStaffUsername = FULL_OFFENCE_DETAILS_REQUEST.victimStaffUsername,
-      victimOtherPersonsName = FULL_OFFENCE_DETAILS_REQUEST.victimOtherPersonsName,
-    )
-    private val FULL_OFFENCE_DETAILS_DB_ENTITY = Offence(
-      offenceCode = FULL_OFFENCE_DETAILS_RESPONSE_DTO.offenceCode,
-      victimPrisonersNumber = FULL_OFFENCE_DETAILS_RESPONSE_DTO.victimPrisonersNumber,
-      victimStaffUsername = FULL_OFFENCE_DETAILS_RESPONSE_DTO.victimStaffUsername,
-      victimOtherPersonsName = FULL_OFFENCE_DETAILS_RESPONSE_DTO.victimOtherPersonsName,
-    )
-
-    private val YOUTH_OFFENCE_DETAILS_REQUEST = OffenceDetailsRequestItem(offenceCode = 2)
-    private val YOUTH_OFFENCE_DETAILS_RESPONSE_DTO = OffenceDetailsDto(
-      offenceCode = YOUTH_OFFENCE_DETAILS_REQUEST.offenceCode,
-      offenceRule = OffenceRuleDetailsDto(
-        paragraphNumber = YOUTH_OFFENCE_CODE_2_PARAGRAPH_NUMBER,
-        paragraphDescription = YOUTH_OFFENCE_CODE_2_PARAGRAPH_DESCRIPTION,
-      )
-    )
-    private val YOUTH_OFFENCE_DETAILS_DB_ENTITY = Offence(
-      offenceCode = YOUTH_OFFENCE_DETAILS_RESPONSE_DTO.offenceCode,
     )
 
     fun incidentRoleDtoWithAllValuesSet(): IncidentRoleDto =
@@ -1483,5 +893,49 @@ class DraftAdjudicationServiceTest {
     assertThat(draftAdjudicationDto.incidentStatement)
       .extracting("statement", "completed")
       .contains("Example statement", false)
+  }
+
+  override fun `throws an entity not found if the draft adjudication for the supplied id does not exists`() {
+    whenever(draftAdjudicationRepository.findById(any())).thenReturn(Optional.empty())
+
+    assertThatThrownBy {
+      draftAdjudicationService.getDraftAdjudicationDetails(1)
+    }.isInstanceOf(EntityNotFoundException::class.java)
+      .hasMessageContaining("DraftAdjudication not found for 1")
+    assertThatThrownBy {
+      draftAdjudicationService.editIncidentRole(1, IncidentRoleRequest("1"), false)
+    }.isInstanceOf(EntityNotFoundException::class.java)
+      .hasMessageContaining("DraftAdjudication not found for 1")
+    assertThatThrownBy {
+      draftAdjudicationService.setIncidentRoleAssociatedPrisoner(
+        1,
+        IncidentRoleAssociatedPrisonerRequest("A1234AA", "A name")
+      )
+    }.isInstanceOf(EntityNotFoundException::class.java)
+      .hasMessageContaining("DraftAdjudication not found for 1")
+    assertThatThrownBy {
+      draftAdjudicationService.addIncidentStatement(1, "test", false)
+    }.isInstanceOf(EntityNotFoundException::class.java)
+      .hasMessageContaining("DraftAdjudication not found for 1")
+
+    assertThatThrownBy {
+      draftAdjudicationService.editIncidentDetails(
+        1,
+        2,
+        DATE_TIME_OF_INCIDENT,
+        DATE_TIME_OF_INCIDENT.plusDays(1),
+      )
+    }.isInstanceOf(EntityNotFoundException::class.java)
+      .hasMessageContaining("DraftAdjudication not found for 1")
+
+    assertThatThrownBy {
+      draftAdjudicationService.editIncidentStatement(1, "new statement", false)
+    }.isInstanceOf(EntityNotFoundException::class.java)
+      .hasMessageContaining("DraftAdjudication not found for 1")
+
+    assertThatThrownBy {
+      draftAdjudicationService.setIncidentApplicableRule(1, false, true)
+    }.isInstanceOf(EntityNotFoundException::class.java)
+      .hasMessageContaining("DraftAdjudication not found for 1")
   }
 }
